Move preserving history

+++ runtime/src/main/kotlin/de/fatox/meta/camera/Gizmo.kt
@@ -12,13 +12,15 @@ import de.fatox.meta.entity.Meta3DEntity
  * Created by Frotty on 19.03.2017.
  */
 class Gizmo {
-    fun showFor(entity: Meta3DEntity?) {
-        val modelBuilder = ModelBuilder()
-        val material = Material(ColorAttribute.createDiffuse(Color.GREEN))
-        val attributes =
-            (VertexAttributes.Usage.Position or VertexAttributes.Usage.Normal or VertexAttributes.Usage.ColorUnpacked).toLong()
-        val xArrow = modelBuilder.createArrow(Vector3.Zero, Vector3.X, material, attributes)
-        val yArrow = modelBuilder.createArrow(Vector3.Zero, Vector3.Y, material, attributes)
-        val zArrow = modelBuilder.createArrow(Vector3.Zero, Vector3.Z, material, attributes)
-    }
+	fun showFor(entity: Meta3DEntity?) {
+		val modelBuilder = ModelBuilder()
+		val material = Material(ColorAttribute.createDiffuse(Color.GREEN))
+		val attributes =
+			(VertexAttributes.Usage.Position or
+				VertexAttributes.Usage.Normal or
+				VertexAttributes.Usage.ColorUnpacked.toLong().toInt()).toLong()
+		val xArrow = modelBuilder.createArrow(Vector3.Zero, Vector3.X, material, attributes)
+		val yArrow = modelBuilder.createArrow(Vector3.Zero, Vector3.Y, material, attributes)
+		val zArrow = modelBuilder.createArrow(Vector3.Zero, Vector3.Z, material, attributes)
+	}
 }
